add a tutorial and automatic execution

+++ tutorials/mock_vasp.py
@@ -10,6 +10,7 @@ All rights reserved.
 
 """
 
+
 import contextlib
 import os
 import shutil
@@ -19,15 +20,15 @@ from pathlib import Path
 
 from pytest import MonkeyPatch
 
-from atomate2.utils.testing.lobster import monkeypatch_lobster
+from atomate2.utils.testing.vasp import monkeypatch_vasp
 
-TEST_ROOT = Path(__file__).parent.parent / "tests"
+TEST_ROOT = Path(__file__).parent.parent/ "tests"
 TEST_DIR = TEST_ROOT / "test_data"
 
 
 @contextlib.contextmanager
-def mock_lobster(ref_paths: dict) -> Generator:
-    """Mock LOBSTER functions.
+def mock_vasp(ref_paths: dict, clean_folders=True) -> Generator:
+    """Mock VASP functions.
 
     Parameters
     ----------
@@ -37,14 +38,14 @@ def mock_lobster(ref_paths: dict) -> Generator:
     ------
         function: A function that mocks calls to VASP.
     """
-    for mf in monkeypatch_lobster(MonkeyPatch(), TEST_DIR / "lobster"):
-        fake_run_lobster_kwargs = {k: {"check_lobster_inputs": ()} for k in ref_paths}
-
+    for mf in monkeypatch_vasp(MonkeyPatch(), TEST_DIR ):
+        fake_run_vasp_kwargs = {k: {"check_inputs": ()} for k in ref_paths}
         old_cwd = os.getcwd()
         new_path = tempfile.mkdtemp()
         os.chdir(new_path)
         try:
-            yield mf(ref_paths, fake_run_lobster_kwargs=fake_run_lobster_kwargs)
+            yield mf(ref_paths, fake_run_vasp_kwargs)
         finally:
             os.chdir(old_cwd)
-            shutil.rmtree(new_path)
+            if clean_folders:
+                shutil.rmtree(new_path)
